+++ libs/sdk-core/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Breaking changes
 
 - Remove support for visual selection labelling
+- Remove support for `keywordset`, `datetime` and `layout` field types
 
 # 1.16.0 (2024-05-29)